Use components logic to draw app
