+++ src/components/about.js
@@ -1,0 +1,9 @@
+const aboutContent = () => `
+  <div>
+    <h1>
+      Here we go
+    </h1>
+  </div>
+`;
+
+export default aboutContent;
